Migrate to MVVM + Material3

+++ app/src/main/java/tech/salroid/filmy/data/local/model/CrewMovie.kt
@@ -38,7 +38,7 @@ data class CrewMovie(
     var video: Boolean? = null,
 
     @SerializedName("vote_average")
-    var voteAverage: Int? = null,
+    var voteAverage: Double? = null,
 
     @SerializedName("vote_count")
     var voteCount: Int? = null,
